--- a/ci/no-xarray.yml
+++ b/ci/no-xarray.yml
@@ -6,6 +6,7 @@ dependencies:
   - netcdf4
   - pandas
   - numpy>=1.22
+  - scipy
   - pip
   - pytest
   - pytest-cov