Closes #95 now that `find_group_cohorts` treats labels as factorized, and `-1` as the NaN sentinel value.

We could go faster with the numpy version but it will blow memory for large enough inputs

-----

sparse bitmask:
```
| Before [19db5b3b] <v0.8.2>   | After [f6af954f] <speedup-cohorts>   |   Ratio | Benchmark (Parameter)                                   |
|------------------------------|--------------------------------------|---------|---------------------------------------------------------|
| 9.59±0.1ms                   | 11.1±0.2ms                           |    1.16 | cohorts.ERA5MonthHourRechunked.time_graph_construct     |
| 172±2ms                      | 134±1ms                              |    0.78 | cohorts.NWMMidwest.time_graph_construct                 |
| 3.15±0.04ms                  | 2.09±0.06ms                          |    0.66 | cohorts.ERA5Google.time_find_group_cohorts              |
| 1.02±0.01ms                  | 550±20μs                             |    0.54 | cohorts.PerfectMonthlyRechunked.time_find_group_cohorts |
| 1.02±0.03ms                  | 530±10μs                             |    0.52 | cohorts.PerfectMonthly.time_find_group_cohorts          |
| 25.3±0.5ms                   | 7.79±0.1ms                           |    0.31 | cohorts.ERA5DayOfYear.time_find_group_cohorts           |
| 7.67±0.1ms                   | 2.20±0.02ms                          |    0.29 | cohorts.ERA5MonthHour.time_find_group_cohorts           |
| 8.16±0.2ms                   | 2.40±0.03ms                          |    0.29 | cohorts.ERA5MonthHourRechunked.time_find_group_cohorts  |
| 70.0±0.5ms                   | 6.74±0.1ms                           |    0.1  | cohorts.ERA5DayOfYearRechunked.time_find_group_cohorts  |
| 233±4ms                      | 19.0±0.6ms                           |    0.08 | cohorts.NWMMidwest.time_find_group_cohorts              |
```


numpy bitmask:

```
| Before [19db5b3b] <v0.8.2>   | After [dd93c1b3] <speedup-cohorts>   |   Ratio | Benchmark (Parameter)                                   |
|------------------------------|--------------------------------------|---------|---------------------------------------------------------|
| 166±0.7ms                    | 143±0.8ms                            |    0.86 | cohorts.NWMMidwest.time_graph_construct                 |
| 3.23±0.02ms                  | 1.23±0ms                             |    0.38 | cohorts.ERA5Google.time_find_group_cohorts              |
| 24.8±0.09ms                  | 7.46±0.02ms                          |    0.3  | cohorts.ERA5DayOfYear.time_find_group_cohorts           |
| 1.02±0ms                     | 293±0.8μs                            |    0.29 | cohorts.PerfectMonthly.time_find_group_cohorts          |
| 1.02±0ms                     | 293±2μs                              |    0.29 | cohorts.PerfectMonthlyRechunked.time_find_group_cohorts |
| 7.78±0.03ms                  | 1.71±0.01ms                          |    0.22 | cohorts.ERA5MonthHourRechunked.time_find_group_cohorts  |
| 214±1ms                      | 47.3±0.5ms                           |    0.22 | cohorts.NWMMidwest.time_find_group_cohorts              |
| 7.50±0.04ms                  | 1.60±0ms                             |    0.21 | cohorts.ERA5MonthHour.time_find_group_cohorts           |
| 64.9±0.07ms                  | 6.25±0.01ms                          |    0.1  | cohorts.ERA5DayOfYearRechunked.time_find_group_cohorts  |
```